Create descion_trees.py
Ginni Impurity is like the confusion level, or like impurity in the nodes descions.

the Gini impurity will be 0. The Gini impurity is highest when the classes are evenly distributed



split_data():
here we are like dividing our data set into 2 conditions yes or no based on thresholds, so this might be a node in which left node has result of those who are within threshold. and right outside it.


best_split():
so it means that if i take this threshold data, then this is the split its making in our data, like 0.8 goes on yes and 0.2 goes on no. so it means that the confusion is there still but its better then 0.50

intuition:

intution is that dude for this data im gonna go thru all the features, and then check which threshold of which feature had the least confusion, i.e ginni index.

so take that and voila we choose this and run this on the trees recursive